Moved default_database_schema to context implementation
