+++ crates/modelardb_common/src/types.rs
@@ -131,11 +131,12 @@ impl fmt::Display for ServerMode {
 }
 
 /// The different possible modes that a ModelarDB server can be deployed in, assigned when the
-/// server is started. MultiNode includes the key received from the manager when registering.
+/// server is started. MultiNode includes the url of the manager and the key received from the
+/// manager when registering.
 #[derive(PartialEq, Eq)]
 pub enum ClusterMode {
     SingleNode,
-    MultiNode(String),
+    MultiNode((String, String)),
 }
 
 #[cfg(test)]
